Add the main folders for the project file structure
