+++ backend/run.py
@@ -1,0 +1,1 @@
+# В этом файле прописываю инструкции для systemd, чтобы быстро развернуть бота на сервере в режиме автозапуска.
